feat: replaced express API server with Bun API server
