+++ tee-apps/sealed-bid-auction/index.ts
@@ -1,31 +1,25 @@
-import {SealedBidAuctionHttpServer} from "./src/api/http/SealedBidAuctionHttpServer.ts";
-import {SealedBidPriceWebsocketServer} from "./src/api/ws/SealedBidPriceWebsocketServer.ts";
+import {SealedBidAuctionApiServer} from "./src/api/SealedBidAuctionApiServer.ts";
 
-const httpServer = new SealedBidAuctionHttpServer();
-const wsServer = new SealedBidPriceWebsocketServer();
+const httpServer = new SealedBidAuctionApiServer();
 
 async function main() {
     await httpServer.start(3010);
-    await wsServer.start(3011);
 }
 
 main()
     .then()
-    .catch((error) => {
-        httpServer.stop();
-        wsServer.stop();
+    .catch(async (error) => {
+        await httpServer.stop();
         console.error("", error);
         process.exit(1);
     });
 
-process.on("SIGINT", () => {
-    httpServer.stop();
-    wsServer.stop();
+process.on("SIGINT", async () => {
+    await httpServer.stop();
     process.exit(0);
 });
 
-process.on("SIGTERM", () => {
-    httpServer.stop();
-    wsServer.stop();
+process.on("SIGTERM", async () => {
+    await httpServer.stop();
     process.exit(0);
 });
