+++ tee-apps/sealed-bid-auction/package.json
@@ -10,13 +10,7 @@
     "typescript": "^5"
   },
   "dependencies": {
-    "@types/cors": "^2.8.19",
-    "@types/express": "^5.0.3",
-    "@types/ws": "^8.18.1",
-    "cors": "^2.8.5",
-    "express": "^5.1.0",
-    "http": "^0.0.1-security",
-    "winston": "^3.17.0",
-    "ws": "^8.18.3"
+    "bun": "^1.2.19",
+    "winston": "^3.17.0"
   }
 }
