+++ tee-apps/sealed-bid-auction/src/api/SealedBidAuctionController.ts
@@ -1,0 +1,7 @@
+import type {BunRequest} from "bun";
+
+export class SealedBidAuctionController {
+    async auction(req: BunRequest): Promise<Response> {
+        return new Response("Not Implemented", {status: 501});
+    }
+}
