--- tee-apps/sealed-bid-auction/src/api/http/SealedBidAuctionController.ts
@@ -1,8 +1,0 @@
-import type {Request, Response} from "express";
-
-export class SealedBidAuctionController {
-
-    async open(req: Request, res: Response): Promise<void> {
-        res.status(501);
-    }
-}
